requirements.txt
streamlit>=1.28.0
plotly>=5.18.0
bcrypt>=4.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.18.0
